requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
beanie==1.25.0
    # via -r requirements.in
broadcaster==0.2.0
    # via -r requirements.in
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   beanie
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.6.1
    # via
    #   email-validator
    #   pymongo
email-validator==2.1.1
    # via pydantic
fastapi==0.110.0
    # via strawberry-graphql
graphql-core==3.2.3
    # via strawberry-graphql
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.4
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via -r requirements.in
idna==3.6
    # via
    #   anyio
    #   email-validator
    #   httpx
lazy-model==0.2.0
    # via beanie
motor==3.3.2
    # via beanie
pydantic[email]==2.6.3
    # via
    #   -r requirements.in
    #   beanie
    #   fastapi
    #   lazy-model
pydantic-core==2.16.3
    # via pydantic
pyjwt==2.8.0
    # via -r requirements.in
pymongo==4.6.2
    # via motor
python-dateutil==2.9.0.post0
    # via strawberry-graphql
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.9
    # via strawberry-graphql
pyyaml==6.0.1
    # via uvicorn
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.36.3
    # via fastapi
strawberry-graphql[fastapi]==0.220.0
    # via -r requirements.in
toml==0.10.2
    # via beanie
typing-extensions==4.10.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   strawberry-graphql
uvicorn[standard]==0.28.0
    # via -r requirements.in
validators==0.22.0
    # via -r requirements.in
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
